pyproject.toml
[project]
name = "tidb-vector-bench"
version = "0.1.0"
description = "Benchmark tool for TiDB Vector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "h5py>=3.13.0",
    "matplotlib>=3.10.3",
    "numpy>=1.24.0",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "ruff>=0.11.10",
    "sqlalchemy>=2.0.41",
    "tqdm>=4.65.0",
    "mysql-connector-python>=8.0.0",
    "psutil>=5.9.0",
    "pyyaml",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tidb_vector_bench"]

[tool.ruff]
line-length = 100
target-version = "py39"
